project: Integrate native cross-process IPC version of obs-studio-node
The module obs-studio-node now consists of obs-studio-client, the client side module, and obs-studio-server, a process that runs outside of Streamlabs OBS. The server has its own crash handler and uses a custom native high performance IPC solution that is drastically faster than the electron IPC and also uses much less IO for complex and binary data. Additionally, the main v8 loop is only blocked if there is an actual call to native code that requires blocking.